requirements.txt
awscli==1.27.144
botocore==1.29.144
cffi==1.15.1
colorama==0.4.4
contourpy==1.1.0
cycler==0.11.0
distlib==0.3.6
docutils==0.16
filelock==3.12.2
fonttools==4.40.0
jmespath==1.0.1
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.25.0
packaging==23.1
Pillow==9.5.0
platformdirs==3.6.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pyparsing==3.1.0
python-dateutil==2.8.2
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16
virtualenv==20.23.1
